Add break option #1508
