+++ README.md
@@ -292,6 +292,7 @@ The following conceptual topics exist in the `PSRule` module:
   - [Configuration](https://aka.ms/ps-rule/options#configuration)
   - [Convention.Include](https://aka.ms/ps-rule/options#conventioninclude)
   - [Execution.AliasReference](https://aka.ms/ps-rule/options#executionaliasreference)
+  - [Execution.Break](https://aka.ms/ps-rule/options#executionbreak)
   - [Execution.DuplicateResourceId](https://aka.ms/ps-rule/options#executionduplicateresourceid)
   - [Execution.HashAlgorithm](https://aka.ms/ps-rule/options#executionhashalgorithm)
   - [Execution.LanguageMode](https://aka.ms/ps-rule/options#executionlanguagemode)
@@ -303,6 +304,7 @@ The following conceptual topics exist in the `PSRule` module:
   - [Execution.UnprocessedObject](https://aka.ms/ps-rule/options#executionunprocessedobject)
   - [Include.Module](https://aka.ms/ps-rule/options#includemodule)
   - [Include.Path](https://aka.ms/ps-rule/options#includepath)
+  - [Input.FileObjects](https://aka.ms/ps-rule/options#inputfileobjects)
   - [Input.Format](https://aka.ms/ps-rule/options#inputformat)
   - [Input.IgnoreGitPath](https://aka.ms/ps-rule/options#inputignoregitpath)
   - [Input.IgnoreObjectSource](https://aka.ms/ps-rule/options#inputignoreobjectsource)
